fix and clean up `PathwaysWebView`
