+++ dbux-common-node/src/util/timeUtil.js
@@ -1,0 +1,6 @@
+export function formatTime(millis) {
+  let seconds = Math.floor(millis / 1000);
+  const minutes = Math.floor(seconds / 60);
+  seconds = seconds - minutes * 60;
+  return `${minutes.toString().padStart(2, '0')}:${seconds.toString().padStart(2, '0')}`;
+}
